### 2019-01-27 #41 @whedon
[PRE REVIEW]: Tutorial for Developing Image-Guided Intervention Modules Using 3D Slicer
**Submitting author:** @lgroves6 (<a href="http://orcid.org/0000-0002-9654-9190">Leah Groves</a>)
**Repository:** <a href="https://github.com/lgroves6/IGTTutorialWebsite" target ="_blank">https://github.com/lgroves6/IGTTutorialWebsite</a>
**Version:** v1.0.0
