### 2019-01-28 #41 @juanklopper
@lgroves6 thank you for your submission.  I am a couple of hours late to reply, but it's been a busy weekend of gunshot violence here in Cape Town and I have been operating 'till early morning every day.  No 3D image guidence here.  Just my two hands and a scalpel. :)

I'd love to learn more about your project.  I note that @labarba has mentioned that a README file is a good place to start.  Could you please populate that for me.  I will study your PDF submission (and the website) as soon as I can, then we can start working on the paper and get some reviewers assigned.

Juan.
