### 2019-01-30 #41 @lgroves6
Hello Juan,

Sorry for the late reply, I hope this e-mail find you well. I hope the gun violence has since settled down. I have populated the ReadMe file for you to review.

Please let me know what the next steps are.

Regards,

Leah
