### 2019-02-03 #41 @juanklopper
@lgroves6 from what I can see, it is an _exceptional_ (static) instructional resource.  I need to hear from @labarba, though, how she sees it fit into our developing ideas around the type of resource we accept for publication.  This is a journal for publishing computational learning modules and educational software according to our website.  A good start therefor would probably be to refer to the work as a learning module instead of a blog as is in the current manuscript (and on the web).
@labarba what are your thoughts?
