### 2019-02-03 #41 @lgroves6
Hello again,

I am confused as to why you are unable to view the README file. Please try to view it through this link https://github.com/lgroves6/Tutorial-for-developing-image-guided-intervention-modules-using-3D-slicer/blob/master/README.md.

Let me know if this works.

Regards,

Leah
