### 2019-02-04 #41 @labarba
I agree with @juanklopper that the use of the term "blog" is confusing. A blog is a rolling sequence of dated posts, often written in personal tones. I don't see that you have dated posts on this webpage. Why call it a blog?
