+++ README.md
@@ -1,1 +1,5 @@
 # IGTTutorialWebsite
+
+This repository contains the source code and a markdown file description of a blog that was developed as a learning-module for the 3D Slicer [@fedorov20123d] architecture and module development for image-guided therapy (IGT) applications. This blog is designed as an interactive wedbsite such that users can scroll over and click on difference components within the blog. In addition to providing the first top-to-bottom description of 3D Slicer architecture, this blog covers the components required to develop an IGT module in 3D Slicer. These components include (1) the PlusServer within the Plus Toolkit [@Lasso2014a], a platform for data capture and broadcasting, and (2) the 3D SlicerIGT Extension [@ungi2016open], a platform for receiving the broadcasted data in 3D Slicer for further processing and visualization. This blog also has examples and video demonstrations for users to follow. By the end of this tutorial the user will understand how 3D Slicer development is structured, how to connect external sources to slicer and how to use these sources to create image-guided modules.  
+
+This blog should be followed in the order of the title page or the order of the table of contents as it is organized that the information progresses in this order. Users could contribute through creating additional video demos to include in the demonstrations protions of the website. 
