+++ paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-This paper presents an overview of a blog (http://computerassistedsurgery.robarts.ca) that was developed as a learning-module for the 3D Slicer [@fedorov20123d] architecture and module development for image-guided therapy (IGT) applications. Image-guided therapies are medical procedures that use computer-based systems that provide navigation facilities and supplemental procedure information to help the physician precisely visualize and target the surgical site. This learning-module focuses on the development of 3D Slicer IGT scripted modules. In addition to providing the first top-to-bottom description of 3D Slicer architecture, this blog covers the components required to develop an IGT module in 3D Slicer. These components include (1) the PlusServer within the Plus Toolkit [@Lasso2014a], a platform for data capture and broadcasting, and (2) the 3D SlicerIGT Extension [@ungi2016open], a platform for receiving the broadcasted data in 3D Slicer for further processing and visualization. This information is presented as an interactive blog in which examples, demonstrations, and sample codes are provided to facilitate implementation of an IGT module for a novice developer.
+This paper presents an overview of a website (http://computerassistedsurgery.robarts.ca) that was developed as a learning-module for the 3D Slicer [@fedorov20123d] architecture and module development for image-guided therapy (IGT) applications. Image-guided therapies are medical procedures that use computer-based systems that provide navigation facilities and supplemental procedure information to help the physician precisely visualize and target the surgical site. This learning-module focuses on the development of 3D Slicer IGT scripted modules. In addition to providing the first top-to-bottom description of 3D Slicer architecture, this module covers the components required to develop an IGT module in 3D Slicer. These components include (1) the PlusServer within the Plus Toolkit [@Lasso2014a], a platform for data capture and broadcasting, and (2) the 3D SlicerIGT Extension [@ungi2016open], a platform for receiving the broadcasted data in 3D Slicer for further processing and visualization. This information is presented as an interactive module in which examples, demonstrations, and sample codes are provided to facilitate implementation of an IGT module for a novice developer.
 
 # Statement of Need
 
