### 2019-02-04 #41 @lgroves6
Hello all,

I have updated the proper README file, as I didn’t realize I had updated the wrong directory. It is now within the proper directory and can be found at this link:
https://github.com/lgroves6/IGTTutorialWebsite/blob/master/paper.md. I have also removed the work blog from the README and paper.md files. You are correct that the term blog is perhaps not the best vocabulary, so we are very happy to term it a learning module.

Please let me know what else you need from me or any changes that should be made.

Thank you for getting back to us and being thorough with your assessment.

Regards,

Leah Groves
