### 2019-02-14 #41 @juanklopper
Thanks @lgroves6 I am happy with those changes.  Perhaps it is time to consider approaching reviewers.  You are free to suggest potential experts.  @labarba do you have anyone in mind that we may approach?
