### 2019-02-19 #41 @lgroves6
Hello again,

I have reach out to a few experts who are interested in being potential reviewers. They are all experts as they are some of the major 3D Slicer developers.

Reviewer information:
 Jean-Christophe Fillion-Robin (@jcfr, jcfr@kitware.com)
  Andrey Fedorov (@fedorov, fedorov@bwh.harvard.edu)
  Matt M (@thewtex, matt.mccormick@kitware.com)

Please let me know if there is any additional information that you need, or anything else I need to do within the review process.
