### 2019-02-26 #46 @labarba
@lgroves6 — At any point, after you make changes to the paper, you can run the command `@whedon generate pdf` to get a new pre-print here.
