### 2019-02-27 #46 @fedorov
@labarba can you confirm this submission is within the scope of this publishing platform?

Reading from the scope of it here http://jose.theoj.org/about

> **What do you mean by "open-source educational materials"?**
>
> Examples include Jupyter notebooks or plaintext/markup language documents like LaTeX, R Markdown, and ReST for course/lesson content and associated notes, with embedded or associated code snippets/programs.
>
>We do not mean openly available slides, lecture notes, or YouTube videos, though these may be acceptable as supplementary materials. In addition, course syllabi by themselves are not suitable for submission (Syllabus may be more appropriate).
>
>**tl;dr**: your course or lesson content must contain or use code to teach. We are not focused exclusively on learning to code, but coding to learn.

The present submission is a web site with the instructions, YouTube videos, etc.

As I was browsing over other issues in this repo, I also came across an earlier submission (#25), which looked interesting to me, so I read through the discussion thread. It is somewhat similar, since it is a collection of non-code-like educational materials, where you @labarba specifically seem to imply that it is out of scope of this platform, see https://github.com/openjournals/jose-reviews/issues/25#issuecomment-462321406.

Please advice.
