### 2019-03-21 #46 @labarba
Indeed, if the materials submitted for review consist on multi-media instructional resources and a website, that is not what JOSE is looking to publish. We are looking for re-usable and complete learning modules that use "coding to learn" (a.k.a. computational thinking). Any videos or slide decks are only considered supplementary to a JOSE publication.

JOSE publishes tutorials with programming as a major component, fully narrated and usable for self-learners or other instructors.
