### 2019-03-22 #46 @labarba
@juanklopper — I see that in the [Pre-Review issue](https://github.com/openjournals/jose-reviews/issues/41#issuecomment-460060385) we already had an exchange about the type of material submitted here, and if it was within scope for JOSE. Nevertheless, we started the review. @fedorov now raises again the in-scope question, having explored the materials further.

I had a bit of trouble finding the [actual website](http://computerassistedsurgery.robarts.ca), as the [submission repository](https://github.com/lgroves6/IGTTutorialWebsite) has not link to it, although it is in the paper. Exploring a bit more now, I see a narrative, with figures and screenshots of an application. The material doesn't use programming, it looks like, although it includes some demos for [developing scripted modules](http://computerassistedsurgery.robarts.ca/create-module/#devmodes)  for the 3D slicer.

After having this deeper look into the materials presented, I do think this submission is out of scope for JOSE. What do you think, @juanklopper ?
