### 2019-03-30 #46 @labarba
👋   @lgroves6 — Thank you for submitting your materials to JOSE. After looking into it further, we determined that this type of content is not in-scope for the journal, as we are targeting learning modules that use programming in some form to support the learning. I am now going to apply the "withdrawn" tag and close the issue.
